pyproject.toml
[project]
name = "meeting-transcript"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.34",
    "langgraph>=0.6.8",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
]
